lesson5/pyproject.toml
[project]
name = "lesson5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio>=4.0.0",
    "dotenv>=0.9.9",
    "python-telegram-bot>=22.5",
]
